% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{createDerivedFtF}
\alias{createDerivedFtF}
\title{Creates useful derived variables and returns dt}
\usage{
createDerivedFtF(dt)
}
\description{
\code{createDerivedFtF}:
}
\details{
sets proper dates and times for R;
   guesses 'home' by looking at location from 01:00 - 04:00. A future version might try geocoding other
   locations (e.g public charging points) via lat/long look-up

adds new derived variables to the passed data.table and returns it
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
