% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftf.R
\name{createSafeFtF}
\alias{createSafeFtF}
\title{Remove disclosive variables}
\usage{
createSafeFtF(dt)
}
\description{
\code{createSafeFtF} creates a new uhnique EV ID byhashing the `Reg No` and then removes the following disclosive variables:
 `Reg No`
 `Latitude`
 `Longitude`
 `Course (deg)`
}
\details{
These should be removed prior to any data sharing.
}
\author{
Ben Anderson, \email{b.anderson@soton.ac.uk} (original)
}
